ai-interview-platform/backend/requirements.txt
fastapi
uvicorn
scikit-learn
spacy
pdfplumber
sentence-transformers
reportlab
pydantic
pymongo
python-dotenv
bcrypt

